funciones calloc y strdup

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: kcarrero <kcarrero@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/20 23:04:29 by kcarrero          #+#    #+#             */
-/*   Updated: 2025/04/20 23:35:03 by kcarrero         ###   ########.fr       */
+/*   Updated: 2025/04/21 11:39:17 by kcarrero         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_atoi(const char *str)
 {
 	int	i;
 	int	sign;
-	long long result;
+	int	result;
 
 	i = 0;
 	sign = 1;
@@ -34,5 +34,5 @@ int	ft_atoi(const char *str)
 		result = result * 10 + (str[i] - '0');
 		i++;
 	}
-	return ((int)(result * sign));
+	return ((result * sign));
 }
